Change memory layout (Paging, Context etc.)
Context - Skeleton for context switching in the future
Kernel - Stays at 2MB offset (size: 2MB)
Page Frame Allocator - After Kernel (size: 2MB). Each context gets his own Page Frame Allocator.
Paging change - After Page Frame Allocator (size: 6MB (reserved space for future tables)).
init_paging - Added additional 2MB of allocation (for Paging structures)

Context switching to Kernel:
Bootloader loads 6MB of pages (2MB - Kernel, 2MB - Page Frame Allocator, 2MB - Paging Tables)
Kernel new PML4 allocates 6MB for Paging Tables (Reserved space).
Kernel new PML4 allocates important Kernel Pages and Page Frame Allocator.
(Total of 10MB for Kernel + Page Frame Allocator + Paging Tables)
Kernel Switches to Kernel Context (new PML4_KERNEL).

+++ libc/include/stdlib.h
@@ -17,20 +17,20 @@ void abort(void);
 /**
  * @brief Convert an integer to a string
  * 
- * @param int number
+ * @param int64_t number
  * @param char* str
  * @param int base (e.g.: 10, 16, 2)
  * 
  * @return char* str
  */
-char* itoa(int, char*, int);
+char* itoa(int64_t, char*, int);
 /**
  * @brief Convert string to int
  * 
  * @param nptr 
  * @return int 
  */
-int atoi(const char *nptr);
+uint64_t atoi(const char *nptr);
 
 #ifdef __cplusplus
 }
